+++ packages/atomic/src/components/commerce/atomic-commerce-interface/atomic-commerce-interface.tsx
@@ -1,11 +1,9 @@
 import {
-  PlatformEnvironment,
   LogLevel,
   Search,
   Unsubscribe,
   UrlManager,
   buildSearch,
-  getOrganizationEndpoints as getOrganizationEndpointsHeadless,
   updateQuery,
   CommerceEngine,
   CommerceEngineConfiguration,
@@ -14,6 +12,9 @@ import {
   ProductListing,
   Context,
   buildContext,
+  Summary,
+  SearchSummaryState,
+  ProductListingSummaryState,
 } from '@coveo/headless/commerce';
 import {
   Component,
@@ -38,10 +39,15 @@ import {
   BaseAtomicInterface,
   CommonAtomicInterfaceHelper,
 } from '../../common/interface/interface-common';
+import {
+  errorSelector,
+  firstSearchExecutedSelector,
+  noProductsSelector,
+} from '../atomic-commerce-layout/commerce-layout';
 import {getAnalyticsConfig} from './analytics-config';
 import {AtomicCommerceStore, createAtomicCommerceStore} from './store';
 
-const FirstSearchExecutedFlag = 'firstSearchExecuted';
+const FirstRequestExecutedFlag = 'firstRequestExecuted';
 
 export type CommerceInitializationOptions = CommerceEngineConfiguration;
 export type CommerceBindings = CommonBindings<
@@ -68,10 +74,11 @@ export class AtomicCommerceInterface
   implements BaseAtomicInterface<CommerceEngine>
 {
   private urlManager!: UrlManager;
-  private searchStatus!: Search | ProductListing;
+  private searchOrListing!: Search | ProductListing;
+  private summary!: Summary<SearchSummaryState | ProductListingSummaryState>;
   private context!: Context;
   private unsubscribeUrlManager: Unsubscribe = () => {};
-  private unsubscribeSearchStatus: Unsubscribe = () => {};
+  private unsubscribeSummary: Unsubscribe = () => {};
   private initialized = false;
   private store: AtomicCommerceStore;
   private commonInterfaceHelper: CommonAtomicInterfaceHelper<CommerceEngine>;
@@ -106,8 +113,10 @@ export class AtomicCommerceInterface
 
   /**
    * the commerce interface language.
+   *
+   * Will default to the value set in the Headless engine context if not provided.
    */
-  @Prop({reflect: true}) public language = 'en';
+  @Prop({reflect: true, mutable: true}) public language?: string;
 
   /**
    * The commerce interface headless engine.
@@ -172,6 +181,7 @@ export class AtomicCommerceInterface
   }
 
   public connectedCallback() {
+    this.store.setLoadingFlag(FirstRequestExecutedFlag);
     this.i18nClone = this.i18n.cloneInstance();
     this.i18n.addResourceBundle = (
       lng: string,
@@ -199,6 +209,9 @@ export class AtomicCommerceInterface
     if (!this.commonInterfaceHelper.engineIsCreated(this.engine)) {
       return;
     }
+    if (!this.language) {
+      return;
+    }
 
     this.context.setLanguage(this.language);
     this.commonInterfaceHelper.onLanguageChange();
@@ -211,7 +224,7 @@ export class AtomicCommerceInterface
 
   public disconnectedCallback() {
     this.unsubscribeUrlManager();
-    this.unsubscribeSearchStatus();
+    this.unsubscribeSummary();
     window.removeEventListener('hashchange', this.onHashChange);
   }
 
@@ -284,18 +297,6 @@ export class AtomicCommerceInterface
     this.engine.executeFirstSearch();
   }
 
-  /**
-   * Returns the unique, organization-specific endpoint(s).
-   * @param {string} organizationId
-   * @param {'prod'|'hipaa'|'staging'|'dev'} [env=Prod]
-   */
-  @Method() public async getOrganizationEndpoints(
-    organizationId: string,
-    env: PlatformEnvironment = 'prod'
-  ) {
-    return getOrganizationEndpointsHeadless(organizationId, env);
-  }
-
   public get bindings(): CommerceBindings {
     return {
       engine: this.engine!,
@@ -353,7 +354,7 @@ export class AtomicCommerceInterface
   }
 
   private initUrlManager() {
-    this.urlManager = this.searchStatus.urlManager({
+    this.urlManager = this.searchOrListing.urlManager({
       initialState: {fragment: this.fragment},
     });
 
@@ -365,17 +366,34 @@ export class AtomicCommerceInterface
   }
 
   private initSearchStatus() {
-    if (this.type === 'product-listing') {
-      this.searchStatus = buildProductListing(this.engine!);
-    } else {
-      this.searchStatus = buildSearch(this.engine!);
-    }
-    this.unsubscribeSearchStatus = this.searchStatus.subscribe(() => {
-      if (
-        !this.searchStatus.state.isLoading &&
-        this.store.hasLoadingFlag(FirstSearchExecutedFlag)
-      ) {
-        this.store.unsetLoadingFlag(FirstSearchExecutedFlag);
+    this.searchOrListing =
+      this.type === 'product-listing'
+        ? buildProductListing(this.engine!)
+        : buildSearch(this.engine!);
+  }
+
+  private initSummary() {
+    this.summary = this.searchOrListing.summary();
+
+    this.unsubscribeSummary = this.summary.subscribe(() => {
+      const {firstRequestExecuted, hasProducts, hasError} = this.summary.state;
+      const hasNoProductsAfterInitialQuery =
+        firstRequestExecuted && !hasError && !hasProducts;
+
+      this.host.classList.toggle(
+        noProductsSelector,
+        hasNoProductsAfterInitialQuery
+      );
+
+      this.host.classList.toggle(errorSelector, hasError);
+
+      this.host.classList.toggle(
+        firstSearchExecutedSelector,
+        firstRequestExecuted
+      );
+
+      if (firstRequestExecuted) {
+        this.store.unsetLoadingFlag(FirstRequestExecutedFlag);
       }
     });
   }
@@ -384,10 +402,16 @@ export class AtomicCommerceInterface
     this.context = buildContext(this.engine!);
   }
 
+  private initLanguage() {
+    if (!this.language) {
+      this.language = this.context.state.language;
+    }
+  }
+
   private updateHash() {
     const newFragment = this.urlManager.state.fragment;
 
-    if (!this.searchStatus.state.isLoading) {
+    if (!this.searchOrListing.state.isLoading) {
       history.replaceState(null, document.title, `#${newFragment}`);
       this.bindings.engine.logger.info(`History replaceState #${newFragment}`);
 
@@ -406,8 +430,10 @@ export class AtomicCommerceInterface
     await this.commonInterfaceHelper.onInitialization(initEngine);
 
     this.initSearchStatus();
+    this.initSummary();
     this.initUrlManager();
     this.initContext();
+    this.initLanguage();
     this.initialized = true;
   }