+++ packages/atomic/src/components/commerce/facets/atomic-commerce-facet/atomic-commerce-facet.tsx
@@ -1,8 +1,9 @@
 import {
+  ProductListingSummaryState,
   RegularFacet,
   RegularFacetState,
-  SearchSummary,
-  ListingSummary,
+  SearchSummaryState,
+  Summary,
 } from '@coveo/headless/commerce';
 import {
   Component,
@@ -61,11 +62,15 @@ export class AtomicCommerceFacet implements InitializableComponent<Bindings> {
   /**
    * The Summary controller instance.
    */
-  @Prop() summary!: SearchSummary | ListingSummary;
+  @Prop() summary!: Summary<SearchSummaryState | ProductListingSummaryState>;
   /**
    * The facet controller instance.
    */
   @Prop() public facet!: RegularFacet;
+  /**
+   * Specifies whether the facet is collapsed.
+   */
+  @Prop({reflect: true, mutable: true}) public isCollapsed = false;
 
   @BindStateToController('facet')
   @State()
@@ -73,7 +78,6 @@ export class AtomicCommerceFacet implements InitializableComponent<Bindings> {
 
   @State() public error!: Error;
 
-  @State() private isCollapsed = false;
   private showLessFocus?: FocusTargetController;
   private showMoreFocus?: FocusTargetController;
   private headerFocus?: FocusTargetController;
@@ -105,12 +109,12 @@ export class AtomicCommerceFacet implements InitializableComponent<Bindings> {
   }
 
   public render() {
-    const {hasError, firstSearchExecuted} = this.summary.state;
+    const {hasError, firstRequestExecuted} = this.summary.state;
     return (
       <FacetGuard
         enabled={true}
         hasError={hasError}
-        firstSearchExecuted={firstSearchExecuted}
+        firstSearchExecuted={firstRequestExecuted}
         hasResults={this.facetState.values.length > 0}
       >
         {