+++ packages/headless/src/controllers/commerce/core/breadcrumb-manager/headless-core-breadcrumb-manager.ts
@@ -252,6 +252,7 @@ export function buildCoreBreadcrumbManager(
 
     deselectAll: () => {
       dispatch(deselectAllBreadcrumbs());
+      dispatch(options.fetchProductsActionCreator());
     },
 
     get state() {