+++ packages/headless/src/features/commerce/common/actions.ts
@@ -1,3 +1,4 @@
+import {getAnalyticsSource} from '../../../api/analytics/analytics-selectors';
 import {getVisitorID} from '../../../api/analytics/coveo-analytics-utils';
 import {SortParam} from '../../../api/commerce/commerce-api-params';
 import {
@@ -59,7 +60,9 @@ export const buildBaseCommerceAPIRequest = async (
     context: {
       user,
       view,
+      capture: state.configuration.analytics.enabled,
       cart: getProductsFromCartState(state.cart),
+      source: getAnalyticsSource(state.configuration.analytics),
     },
     ...effectivePagination(state, slotId),
   };