+++ packages/headless/src/features/commerce/pagination/pagination-slice.ts
@@ -1,4 +1,5 @@
 import {createReducer} from '@reduxjs/toolkit';
+import {deselectAllBreadcrumbs} from '../../breadcrumb/breadcrumb-actions';
 import {deselectAllFacetValues} from '../../facets/facet-set/facet-set-actions';
 import {setContext, setUser, setView} from '../context/context-actions';
 import {toggleSelectCategoryFacetValue} from '../facets/category-facet/category-facet-actions';
@@ -103,6 +104,7 @@ export const paginationReducer = createReducer(
       })
       .addCase(restoreSearchParameters, handleRestoreParameters)
       .addCase(restoreProductListingParameters, handleRestoreParameters)
+      .addCase(deselectAllBreadcrumbs, handlePaginationReset)
       .addCase(deselectAllFacetValues, handlePaginationReset)
       .addCase(toggleSelectFacetValue, handlePaginationReset)
       .addCase(toggleExcludeFacetValue, handlePaginationReset)